requirements.txt
googletrans==3.1.0a0
Pillow
pytesseract
nltk
numpy
networkx
flask
gunicorn
joblib
dill
sklearn
